- transfers adding to favorites from the  FeedFragment to NewsDetailsFragment
