+++ app/src/main/java/com/innoprog/android/feature/feed/newsdetails/di/NewsDetailsModule.kt
@@ -3,6 +3,10 @@ package com.innoprog.android.feature.feed.newsdetails.di
 import androidx.lifecycle.ViewModel
 import com.innoprog.android.di.ViewModelKey
 import com.innoprog.android.feature.feed.newsdetails.presentation.NewsDetailsViewModel
+import com.innoprog.android.feature.feed.newsfeed.data.FavoritesRepositoryImpl
+import com.innoprog.android.feature.feed.newsfeed.domain.FavoritesInteractor
+import com.innoprog.android.feature.feed.newsfeed.domain.FavoritesRepository
+import com.innoprog.android.feature.feed.newsfeed.domain.impl.FavoritesInteractorImpl
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
@@ -13,4 +17,10 @@ interface NewsDetailsModule {
     @ViewModelKey(NewsDetailsViewModel::class)
     @Binds
     fun bindFeedViewModel(impl: NewsDetailsViewModel): ViewModel
+
+    @Binds
+    fun bindFavoritesRepository(favoritesRepositoryImpl: FavoritesRepositoryImpl): FavoritesRepository
+
+    @Binds
+    fun bindFavoritesInteractor(favoritesInteractorImpl: FavoritesInteractorImpl): FavoritesInteractor
 }
