+++ app/src/main/java/com/innoprog/android/feature/feed/newsdetails/presentation/NewsDetailsFragment.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.viewpager2.widget.ViewPager2
+import android.widget.Toast
 import com.innoprog.android.R
 import com.innoprog.android.base.BaseFragment
 import com.innoprog.android.base.BaseViewModel
@@ -12,14 +12,12 @@ import com.innoprog.android.databinding.FragmentNewsDetailsBinding
 import com.innoprog.android.di.AppComponentHolder
 import com.innoprog.android.di.ScreenComponent
 import com.innoprog.android.feature.feed.newsdetails.di.DaggerNewsDetailsComponent
-import com.innoprog.android.feature.feed.newsfeed.presentation.FeedViewModel
 import com.innoprog.android.feature.imagegalleryadapter.ImageGalleryAdapter
 
 class NewsDetailsFragment : BaseFragment<FragmentNewsDetailsBinding, BaseViewModel>() {
 
     override val viewModel by injectViewModel<NewsDetailsViewModel>()
-    private lateinit var viewPager2: ViewPager2
-    private lateinit var adapter: ImageGalleryAdapter
+    private var adapter: ImageGalleryAdapter? = null
 
     override fun diComponent(): ScreenComponent {
         val appComponent = AppComponentHolder.getComponent()
@@ -38,17 +36,27 @@ class NewsDetailsFragment : BaseFragment<FragmentNewsDetailsBinding, BaseViewMod
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        setUiListeners()
+        initImageGallery()
+    }
+
+    private fun setUiListeners() {
         binding.newsTopBar.setLeftIconClickListener {
             viewModel.navigateUp()
         }
 
-        viewPager2 = view.findViewById(R.id.viewPager)
+        binding.newsTopBar.setRightIconClickListener {
+            Toast.makeText(requireContext(), "Добавлено/удалено из избранного", Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    private fun initImageGallery(){
         val images = listOf(
             R.drawable.news_sample,
             R.drawable.course_logo_sample
         )
 
         adapter = ImageGalleryAdapter(images)
-        viewPager2.adapter = adapter
+        binding.viewPager.adapter = adapter
     }
 }
