+++ app/src/main/java/com/innoprog/android/feature/feed/newsfeed/presentation/FeedViewModel.kt
@@ -1,17 +1,8 @@
 package com.innoprog.android.feature.feed.newsfeed.presentation
 
-import androidx.lifecycle.viewModelScope
 import com.innoprog.android.base.BaseViewModel
-import com.innoprog.android.feature.feed.newsfeed.domain.FavoritesInteractor
-import com.innoprog.android.feature.feed.newsfeed.domain.models.News
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-class FeedViewModel @Inject constructor(private val favoritesInteractor: FavoritesInteractor) :
+class FeedViewModel @Inject constructor() :
     BaseViewModel() {
-    fun onFavoriteClicked(news: News) {
-        viewModelScope.launch {
-            favoritesInteractor.insertNewsToFavorites(news)
-        }
-    }
 }
