+++ uikit/src/main/res/drawable/ic_comment_dark.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M6.799,17.803C6.357,18.13 5.73,18.041 5.399,17.604C5.27,17.434 5.2,17.226 5.2,17.013L5.199,14.842H4.6C3.164,14.842 2,13.693 2,12.276V5.566C2,4.149 3.164,3 4.6,3H15.4C16.836,3 18,4.149 18,5.566V12.276C18,13.693 16.836,14.842 15.4,14.842H10.81L6.799,17.803Z"
+      android:fillColor="#637083"/>
+</vector>
